Creating in-sort team-division algorithm
Used algorithm that divides the teams as it reads the experience levels. This occurs by using opposing index values for experienced and inexperienced players. (exp_index moves from 0 up to the number of teams - 1, and inexp_index moves from number of teams - 1 to 0.) This allows for experience to be divided as evenly as possible between the teams  even if the number of players is not evenly divisible by the number of teams OR if the number of experienced/inexperienced players is not divisible by the number of teams.

Also, refractored the print_player function into the player_description function that returns the player description instead of printint it.